+++ filter/bandpass.h
@@ -4,21 +4,21 @@
 #define _FILTER_BANDPASS_H_
 
 struct filter_bandpass_st {
-    double now;
-    double next;
-    double lastfreq;
-    double lastres;
-    double fx;
-    double a0;
-    double last1;
-    double last2;
-    double *input;
-    double *freq;
-    double *res;
+    float now;
+    float next;
+    float lastfreq;
+    float lastres;
+    float fx;
+    float a0;
+    float last1;
+    float last2;
+    float *input;
+    float *freq;
+    float *res;
 };
 
-// !lua:filter_bandpass -> filter_bandpass_make(double *input, double *frequency, double *resonance)
-struct filter_bandpass_st * filter_bandpass_make(double *input, double *frequency, double *resonance);
+// !lua:filter_bandpass -> filter_bandpass_make(float *input, float *frequency, float *resonance)
+struct filter_bandpass_st * filter_bandpass_make(float *input, float *frequency, float *resonance);
 
 #endif
